requirements.txt
# Minimal dependencies for Kraken Subaccount Transfer Tool
#
# Core dependencies only - minimal attack surface
# All versions pinned for reproducibility and security

# HTTP requests to Kraken API
requests>=2.31.0,<3.0.0

# Environment variable management (.env file loading)
python-dotenv>=1.0.0,<2.0.0

# Data display and formatting
pandas>=2.0.0,<3.0.0

# Jupyter notebook
jupyter>=1.0.0,<2.0.0
ipykernel>=6.25.0,<7.0.0
